fix: fix a regression bug of rendering the alert icon by severity
