+++ packages/react/src/alert/AlertIcon.js
@@ -1,6 +1,5 @@
 import React, { forwardRef } from 'react';
 import { Box } from '../box';
-import { Icon } from '../icon';
 import {
   defaultSeverity,
   defaultVariant,
@@ -9,26 +8,8 @@ import {
   useAlertIconStyle,
 } from './styles';
 
-const getIconBySeverity = (severity) => {
-  const iconName = {
-    success: 'success',
-    info: 'info',
-    warning: 'warning-triangle',
-    error: 'error',
-  }[severity];
-
-  if (!iconName) {
-    return null;
-  }
-
-  return (
-    <Icon icon={`${iconName}`} />
-  );
-};
-
 const AlertIcon = forwardRef((
   {
-    icon,
     severity = defaultSeverity,
     variant = defaultVariant,
     ...rest
@@ -37,21 +18,12 @@ const AlertIcon = forwardRef((
 ) => {
   const styleProps = useAlertIconStyle({ variant, severity });
 
-  if (typeof icon === 'string') {
-    icon = (<Icon icon={icon} />);
-  }
-  if (typeof icon === 'undefined') {
-    icon = getIconBySeverity(severity);
-  }
-
   return (
     <Box
       ref={ref}
       {...styleProps}
       {...rest}
-    >
-      {icon}
-    </Box>
+    />
   );
 });
